+++ CHANGELOG.md
@@ -10,6 +10,8 @@ IMPROVEMENTS:
 
 * core: built with Go 1.1.2
 * core: packer help output now loads much faster.
+* builder/virtualbox: guest_additions_url can now use the `Version`
+  variable to get the VirtualBox version. [GH-272]
 * builder/virtualbox: Do not check for VirtualBox as part of template
   validation; only check at execution.
 * builder/vmware: Do not check for VMware as part of template validation;